refactor(profiling): extract functions from lib.rs to wall_time.rs
